Switched to a new way to split the lyrics
The new way tries to fit as many verses into the available 2000 characters as possible. It does this by building up the page 1 verse at a time, until it runs out of space. At which time it adds it to the array of pages and restarts, until it runs out of verses.